изменяет render и конструктор
