+++ src/presenter/itinerary-presenter.js
@@ -1,18 +1,20 @@
 import EventsListView from '../view/trip-events-list-view.js';
-import NewPointVeiw from '../view/addNewPoint-view.js';
+import NewPointView from '../view/add-new-point-view.js';
+
 import { render } from '../render.js';
 
 
 export default class ItineraryPresenter {
   eventListComponent = new EventsListView();
-  newPointComponent = new NewPointVeiw();
+  newPointComponent = new NewPointView();
 
-  constructor({ container }) {
-    this.container = container;
+  constructor({ tripEventsElement }) {
+    this.tripEventsElement = tripEventsElement;
   }
 
   init() {
-    render(this.eventListComponent, this.container);
-    render(this.newPointComponent, this.container);
+    render(this.eventListComponent, this.tripEventsElement);
+    // Render newPointComponent inside the event list
+    render(this.newPointComponent, this.eventListComponent.getElement());
   }
 }
