+++ src/render.js
@@ -12,11 +12,14 @@ function createElement(template) {
   return newElement.firstElementChild;
 }
 
-function render(component, container, place = RenderPosition.BEFOREEND) {
-  container.insertAdjacentElement(place, component.getElement());
-}
-function renderBgn(component, container, place = RenderPosition.AFTERBEGIN) {
+/**
+ * Renders a component into a container at the specified position.
+ * @param {Object} component - The component instance with a getElement() method.
+ * @param {Element} container - The DOM element to render into.
+ * @param {string} place - The position relative to the container (use RenderPosition).
+ */
+function render(component, container, place) {
   container.insertAdjacentElement(place, component.getElement());
 }
 
-export { RenderPosition, createElement, render, renderBgn };
+export { RenderPosition, createElement, render };
